refactor: move function
Function acts on domain-types and so should be in a different file
