+++ src/types/domain-name.ts
@@ -55,3 +55,18 @@ export function isDomainName(s: string): s is DomainName {
   if (segments === null || segments.length === 0) return false;
   return segments.every((segment) => segmentRegex.test(segment));
 }
+
+/**
+ * Detect if the given package name is an internal package
+ * @param name The name of the package
+ */
+export const isInternalPackage = (name: DomainName): boolean => {
+  const internals = [
+    "com.unity.ugui",
+    "com.unity.2d.sprite",
+    "com.unity.2d.tilemap",
+    "com.unity.package-manager-ui",
+    "com.unity.ugui",
+  ];
+  return /com.unity.modules/i.test(name) || internals.includes(name);
+};
