+++ src/utils/pkg-name.ts
@@ -25,18 +25,3 @@ export const splitPkgName = function (pkgName: PkgName): {
  */
 export const atVersion = (name: DomainName, version: PkgVersion): PkgName =>
   `${name}@${version}`;
-
-/**
- * Detect if the given package name is an internal package
- * @param name The name of the package
- */
-export const isInternalPackage = (name: DomainName): boolean => {
-  const internals = [
-    "com.unity.ugui",
-    "com.unity.2d.sprite",
-    "com.unity.2d.tilemap",
-    "com.unity.package-manager-ui",
-    "com.unity.ugui",
-  ];
-  return /com.unity.modules/i.test(name) || internals.includes(name);
-};
